+++ client/src/contract/error.rs
@@ -1,7 +1,9 @@
+use std::borrow::Cow;
 use std::sync::Arc;
 
 use thiserror::Error;
 use tonlib_core::cell::TonCellError;
+use tonlib_core::types::TonHashParseError;
 use tonlib_core::TonAddress;
 
 use crate::client::TonClientError;
@@ -77,7 +79,7 @@ pub enum TonContractError {
     TvmStackParseError {
         method: TonMethodId,
         address: TonAddress,
-        error: StackParseError,
+        error: Box<StackParseError>,
     },
 
     #[error(
@@ -87,22 +89,25 @@ pub enum TonContractError {
     TvmRunError {
         method: TonMethodId,
         address: TonAddress,
-        vm_log: Option<String>,
+        vm_log: Box<Option<String>>,
         exit_code: i32,
-        stack: Vec<TvmStackEntry>,
+        stack: Box<Vec<TvmStackEntry>>,
         missing_library: Option<String>,
         gas_used: i64,
     },
 
     #[error("{0}")]
     CacheError(#[from] Arc<TonContractError>),
+
+    #[error("{0}")]
+    TonLibraryError(#[from] TonLibraryError),
 }
 
 pub trait MapStackError<R> {
     #[allow(clippy::result_large_err)]
     fn map_stack_error(
         self,
-        method: &'static str,
+        method: impl Into<Cow<'static, str>>,
         address: &TonAddress,
     ) -> Result<R, TonContractError>;
 }
@@ -111,7 +116,7 @@ pub trait MapCellError<R> {
     #[allow(clippy::result_large_err)]
     fn map_cell_error(
         self,
-        method: &'static str,
+        method: impl Into<Cow<'static, str>>,
         address: &TonAddress,
     ) -> Result<R, TonContractError>;
 }
@@ -119,27 +124,29 @@ pub trait MapCellError<R> {
 impl<R> MapStackError<R> for Result<R, TvmStackError> {
     fn map_stack_error(
         self,
-        method: &'static str,
+        method: impl Into<Cow<'static, str>>,
         address: &TonAddress,
     ) -> Result<R, TonContractError> {
-        self.map_err(|e| TonContractError::MethodResultStackError {
-            method: method.into(),
-            address: address.clone(),
-            error: e,
-        })
+        self.map_err(
+            |e: TvmStackError| TonContractError::MethodResultStackError {
+                method: TonMethodId::Name(method.into()),
+                address: address.clone(),
+                error: e,
+            },
+        )
     }
 }
 
 impl<R> MapStackError<R> for Result<R, StackParseError> {
     fn map_stack_error(
         self,
-        method: &'static str,
+        method: impl Into<Cow<'static, str>>,
         address: &TonAddress,
     ) -> Result<R, TonContractError> {
         self.map_err(|e| TonContractError::TvmStackParseError {
-            method: method.into(),
+            method: TonMethodId::Name(method.into()),
             address: address.clone(),
-            error: e,
+            error: e.into(),
         })
     }
 }
@@ -147,13 +154,34 @@ impl<R> MapStackError<R> for Result<R, StackParseError> {
 impl<R> MapCellError<R> for Result<R, TonCellError> {
     fn map_cell_error(
         self,
-        method: &'static str,
+        method: impl Into<Cow<'static, str>>,
         address: &TonAddress,
     ) -> Result<R, TonContractError> {
         self.map_err(|e| TonContractError::MethodResultStackError {
-            method: method.into(),
+            method: TonMethodId::Name(method.into()),
             address: address.clone(),
             error: e.into(),
         })
     }
+}
+
+#[derive(Error, Debug)]
+pub enum TonLibraryError {
+    #[error("{0}")]
+    TonClientError(#[from] TonClientError),
+
+    #[error("{0}")]
+    TonCellError(#[from] TonCellError),
+
+    #[error("{0}")]
+    TonHashParseError(#[from] TonHashParseError),
+
+    #[error("Library not found for {0}")]
+    LibraryNotFound(String),
+
+    #[error("Expected exactly one library, but got multiple")]
+    MultipleLibrariesReturned,
+
+    #[error("Getting library by mc_seqno is not supported")]
+    SeqnoNotSupported,
 }